+++ src/app/records/PresentationNodeLeaf.tsx
@@ -1,4 +1,3 @@
-import { orderBy } from 'lodash';
 import Collectible from './Collectible';
 import CollectiblesGrid from './CollectiblesGrid';
 import Craftable from './Craftable';
@@ -14,22 +13,12 @@ export default function PresentationNodeLeaf({
   ownedItemHashes,
   completedRecordsHidden,
   redactedRecordsRevealed,
-  sortRecordProgression,
 }: {
   node: DimPresentationNodeLeaf;
   ownedItemHashes?: Set<number>;
   completedRecordsHidden: boolean;
   redactedRecordsRevealed: boolean;
-  sortRecordProgression: boolean;
 }) {
-  if (!sortRecordProgression && 'childPresentationNodes' in node && 'acquired' in node) {
-    node.childPresentationNodes = orderBy(
-      node.childPresentationNodes!,
-      (entry) => entry['acquired'] / entry['visible'],
-      ['asc']
-    );
-  }
-
   return (
     <>
       {node.collectibles && node.collectibles.length > 0 && (