+++ docs/CHANGELOG.md
@@ -9,7 +9,7 @@
 * `is:inloadout:` searches now autocomplete hashtags in loadout names and descriptions.
 * Fixed plugging Harmonic Resonance mods when using a Strand subclass.
 * You can now drag and drop subclasses into the loadout editor, including from other loadouts.
-* Weapon patterns are now sorted by completion %.
+* Added a toggle button for sorting tabs in the records page.
 
 ## 7.70.0 <span class="changelog-date">(2023-05-28)</span>